fix: Moved recover test to own file
